requirements.txt
streamlit==1.29.0
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
pandas==2.1.3
google-generativeai==0.3.1
python-dotenv==1.0.0
easyocr==1.7.1
numpy==1.26.2
kaggle
